#430: finish the culture widening at its consumers -- 33 int64_t truncations
The culture 64-bit widening was half-landed. Storage and getters went
int64_t because the overflow was live in shipping saves; 33 consumer
sites still took them as `int` and truncated straight back, which is the
saturating, plausible-looking wrong number the widening exists to end,
reintroduced one layer out.

The compiler had been reporting all 33 on every build. Nobody read them.

THE WRITE PATH was the worst of it. CvCity::changeCultureTimes100 takes
int64_t and calls setCultureTimes100(int64_t), but the local between them
was `int`, guarded on MAX_INT and pinned at MAX_INT on overflow -- so
every write clamped city culture to 32 bits regardless of what either
side could carry. Now int64_t throughout, saturating at LLONG_MAX.

OWNERSHIP. CvPlot::findHighestCulturePlayer compared truncated culture
values to pick the highest -- and that feeds cultural ownership, so a
wrapped value could hand a plot to the wrong player. Same shape in
CvPlot's largest-culture scan and in CvPlayer's plot_owners pair.

THREE OVERFLOW WORKAROUNDS DELETED, not preserved. getCulturePercentAnger,
the espionage culture-insert and its AI twin each carried a
`if (culture > MAX_INT/N)` branch that dropped precision to dodge a 32-bit
overflow. The premise is gone, so the branch goes and the precise form
stands -- the magic constant disappearing is the signal the width is now
right. ⚠ The AI one was also WRONG: it scaled iPlotCulture by 1000, then
computed the amount from the UNSCALED getCulture() and multiplied by the
scale anyway. Collapsing it fixes that.

PUBLISHED READS. CyPlot::getCulture and countTotalCulture are live Python
bindings and were truncating; they widen rather than being cut, since
int64_t already crosses this boundary and is already published elsewhere
(getRealPopulation, getFinalExpense).

WHAT STAYS int, deliberately, because the value genuinely is bounded and
the reduce belongs at the point of use ([DEC-fixedpoint-x100]): the
percent-anger ratio, calculateCulturePercent, the two influence per-mille
returns, the turns-left displays, the team culture ratio, and the per-turn
culture RATE vectors -- those are cleared every turn, do not accumulate,
and are serialized as int.

Assert green. int64_t->int narrowings 33 -> 0; total warnings 44 -> 11.

+++ Sources/AI/CvUnitAI.cpp
@@ -15639,7 +15639,7 @@ bool CvUnitAI::AI_greatWork()
 
 	CvPlot* pBestPlot = NULL;
 	{
-		int iBestValue = 0;
+		int64_t iBestValue = 0;
 		foreach_(const CvCity * cityX, GET_PLAYER(getOwner()).cities())
 		{
 			if (cityX->area() == area()
@@ -15648,7 +15648,7 @@ bool CvUnitAI::AI_greatWork()
 			&& GET_PLAYER(getOwner()).AI_plotTargetMissionAIs(cityX->plot(), MISSIONAI_GREAT_WORK, getGroup()) == 0
 			&& generateIgnoreDangerPath(cityX->plot()))
 			{
-				const int iValue =
+				const int64_t iValue =
 					(
 						cityX->AI_calculateCulturePressure(true)
 						-
